[cleanup] Remove obsolete Strategy Architect — debate society is the sole generation path (#1064)
