+++ backend/archimedes/services/__init__.py
@@ -10,7 +10,6 @@ orchestrator.
 # (scripts, notebooks, stale imports) working without changes.
 from archimedes.agents import generation_pipeline as generation_pipeline
 from archimedes.agents import portfolio_agent as portfolio_agent
-from archimedes.agents import strategy_architect as strategy_architect
 from archimedes.agents import strategy_fusion as strategy_fusion
 
 # Re-export so MarketplaceState's dynamic import via
