+++ backend/tests/services/test_strategy_guardrail.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 
 import pytest
-from archimedes.agents.strategy_architect import ArchitectProposal, StrategySelection
+from archimedes.agents.generation_json import ArchitectProposal, StrategySelection
 from archimedes.services.strategy_guardrail import (
     DEFAULT_MAX_STRATEGY_WEIGHT,
     GuardrailResult,
